fixed float->float64 issue

+++ demo/demo2/demo2.go
@@ -31,14 +31,14 @@ func main() {
 		}
 		end = time.Nanoseconds()
 		duration = end - start
-		fmt.Printf("%d %dx%d x %dx%d matrix multiplications in %fs\n", count, h, w, w, h, float(end-start)/1000000000)
+		fmt.Printf("%d %dx%d x %dx%d matrix multiplications in %fs\n", count, h, w, w, h, float64(end-start)/1000000000)
 
 		start = time.Nanoseconds()
 		for i := 0; i < count; i++ {
 			A.TimesDense(B)
 		}
 		end = time.Nanoseconds()
-		fmt.Printf("%d %dx%d x %dx%d dense matrix multiplications in %fs\n", count, h, w, w, h, float(end-start)/1000000000)
+		fmt.Printf("%d %dx%d x %dx%d dense matrix multiplications in %fs\n", count, h, w, w, h, float64(end-start)/1000000000)
 		fmt.Printf("For a ratio of %f.\n", float64(duration)/float64(end-start))
 	}
 
@@ -52,14 +52,14 @@ func main() {
 	}
 	end = time.Nanoseconds()
 	duration = end - start
-	fmt.Printf("%d %dx%d x %dx%d matrix additions in %fs\n", count, h, w, w, h, float(end-start)/1000000000)
+	fmt.Printf("%d %dx%d x %dx%d matrix additions in %fs\n", count, h, w, w, h, float64(end-start)/1000000000)
 
 	start = time.Nanoseconds()
 	for i := 0; i < count; i++ {
 		A.PlusDense(A)
 	}
 	end = time.Nanoseconds()
-	fmt.Printf("%d %dx%d x %dx%d dense matrix additions in %fs\n", count, h, w, w, h, float(end-start)/1000000000)
+	fmt.Printf("%d %dx%d x %dx%d dense matrix additions in %fs\n", count, h, w, w, h, float64(end-start)/1000000000)
 	fmt.Printf("For a ratio of %f.\n", float64(duration)/float64(end-start))
 
 	A = Normals(4, 4)
@@ -71,7 +71,7 @@ func main() {
 		A.ElementMult(B)
 	}
 	end = time.Nanoseconds()
-	fmt.Printf("%d 4x4 matrix element multiplications in %fs\n", count, float(end-start)/1000000000)
+	fmt.Printf("%d 4x4 matrix element multiplications in %fs\n", count, float64(end-start)/1000000000)
 
 	A = Normals(4, 4)
 	B = Normals(4, 4)
@@ -82,7 +82,7 @@ func main() {
 		A.Plus(B)
 	}
 	end = time.Nanoseconds()
-	fmt.Printf("%d 4x4 matrix additions in %fs\n", count, float(end-start)/1000000000)
+	fmt.Printf("%d 4x4 matrix additions in %fs\n", count, float64(end-start)/1000000000)
 
 	A = Normals(6, 6)
 
@@ -92,7 +92,7 @@ func main() {
 		A.Inverse()
 	}
 	end = time.Nanoseconds()
-	fmt.Printf("%d 6x6 matrix inversions in %fs\n", count, float(end-start)/1000000000)
+	fmt.Printf("%d 6x6 matrix inversions in %fs\n", count, float64(end-start)/1000000000)
 
 	A = Normals(6, 6)
 
@@ -102,7 +102,7 @@ func main() {
 		A.Det()
 	}
 	end = time.Nanoseconds()
-	fmt.Printf("%d 6x6 matrix determinants in %fs\n", count, float(end-start)/1000000000)
+	fmt.Printf("%d 6x6 matrix determinants in %fs\n", count, float64(end-start)/1000000000)
 
 	A = Normals(6, 6)
 	Bm, _ := A.Times(A.Transpose())
@@ -114,7 +114,7 @@ func main() {
 		B.Cholesky()
 	}
 	end = time.Nanoseconds()
-	fmt.Printf("%d 6x6 cholesky decompositions in %fs\n", count, float(end-start)/1000000000)
+	fmt.Printf("%d 6x6 cholesky decompositions in %fs\n", count, float64(end-start)/1000000000)
 
 	A = Normals(6, 6)
 
@@ -124,7 +124,7 @@ func main() {
 		A.QR()
 	}
 	end = time.Nanoseconds()
-	fmt.Printf("%d 6x6 QR decompositions in %fs\n", count, float(end-start)/1000000000)
+	fmt.Printf("%d 6x6 QR decompositions in %fs\n", count, float64(end-start)/1000000000)
 
 	A = Normals(6, 6)
 
@@ -134,7 +134,7 @@ func main() {
 		A.Eigen()
 	}
 	end = time.Nanoseconds()
-	fmt.Printf("%d 6x6 Eigenvector decompositions in %fs\n", count, float(end-start)/1000000000)
+	fmt.Printf("%d 6x6 Eigenvector decompositions in %fs\n", count, float64(end-start)/1000000000)
 
 	A = Normals(6, 6)
 
@@ -144,6 +144,6 @@ func main() {
 		A.SVD()
 	}
 	end = time.Nanoseconds()
-	fmt.Printf("%d 6x6 singular value decompositions in %fs\n", count, float(end-start)/1000000000)
+	fmt.Printf("%d 6x6 singular value decompositions in %fs\n", count, float64(end-start)/1000000000)
 
 }
